OLC and mobprogs code (1.71) merged (yeah!). Old mobprogs code removed at all. objprogs re-implemented (for OLC). Some code cleanups in act_printf/act_nprintf/bust_a_prompt. do_wiznet syntax changed.
git-svn-id: file:///home/fjoe/old-repos/svn/husky/trunk/sog@377 1738b366-ca78-dd11-ac90-000e8e03bdb1

+++ src/sog/flags.c
@@ -1,5 +1,5 @@
 /*
- * $Id: flags.c,v 1.3 1998-04-26 17:08:05 efdi Exp $
+ * $Id: flags.c,v 1.4 1998-06-28 04:47:14 fjoe Exp $
  */
 
 /***************************************************************************
@@ -57,7 +57,7 @@ void do_flag(CHAR_DATA *ch, char *argument)
     char arg1[MAX_INPUT_LENGTH],arg2[MAX_INPUT_LENGTH],arg3[MAX_INPUT_LENGTH];
     char word[MAX_INPUT_LENGTH];
     CHAR_DATA *victim;
-    long *flag, old = 0, new = 0, marked = 0, pos;
+    int *flag, old = 0, new = 0, marked = 0, pos;
     char type;
     const struct flag_type *flag_table;
